TMario::TSurfingParams::TSurfingParams: mRoll is TParamRT<f32>, not s32

+++ include/Player/Mario.hpp
@@ -177,7 +177,7 @@ public:
 		TParamRT<f32> mWaistPitch;
 		TParamRT<s16> mWaistRollMax;
 		TParamRT<s16> mWaistPitchMax;
-		TParamRT<s32> mRoll;
+		TParamRT<f32> mRoll;
 		TParamRT<f32> mPitch;
 		TParamRT<s16> mRollMax;
 		TParamRT<s16> mPitchMax;
